pyproject.toml
[project]
name = "sg-ai-jobs"
version = "0.1.0"
description = "Singapore AI Jobs Exposure Treemap — interactive visualization of AI automation risk across ~560 occupations"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1",
    "pandas>=2.2",
    "pydantic>=2.10",
    "anthropic>=0.45",
    "python-dotenv>=1.0",
    "rapidfuzz>=3.10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8.3",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
